H3/Evaluation/requirements.txt
Flask==1.1.2
google-cloud-datastore==2.1.0
google-cloud-storage==1.37.0
cerberus==1.3.2
